splits up the section unit tests into several files
