+++ test/util/testUtil.go
@@ -67,34 +67,6 @@ func CheckMessage(m1, m2 message.Message, t *testing.T) {
 	}
 }
 
-func CheckQuery(q1, q2 *query.Name, t *testing.T) {
-	if q1.Context != q2.Context {
-		t.Error("Query context mismatch")
-	}
-	if q1.Expiration != q2.Expiration {
-		t.Error("Query Expires mismatch")
-	}
-	if q1.Name != q2.Name {
-		t.Error("Query Name mismatch")
-	}
-	if len(q1.Types) != len(q2.Types) {
-		t.Error("Query Type length mismatch")
-	}
-	for i, o1 := range q1.Types {
-		if o1 != q2.Types[i] {
-			t.Errorf("Query Type at position %d mismatch", i)
-		}
-	}
-	if len(q1.Options) != len(q2.Options) {
-		t.Error("Query Option length mismatch")
-	}
-	for i, o1 := range q1.Options {
-		if o1 != q2.Options[i] {
-			t.Errorf("Query Option at position %d mismatch", i)
-		}
-	}
-}
-
 func CheckNotification(n1, n2 *section.Notification, t *testing.T) {
 	if n1.Type != n2.Type {
 		t.Error("Notification Type mismatch")
